test(evals): cover no_component_paywall in CI test path + fix evaluator count
Rebase onto main brought 6 evaluators/rules (redo_of_shipped_capability +
no_component_paywall); bump the count assertion 5->6. Port no_component_paywall
shape + rule structural tests into pipeline/tests/ (the CI testpaths) and drop
the duplicate under pipeline/evals/ that sat outside testpaths and never ran in CI.
