+++ pipeline/tests/test_setup_langfuse_evaluators.py
@@ -13,6 +13,7 @@ from evals import setup_langfuse_evaluators as mod  # noqa: E402
 from evals.setup_langfuse_evaluators import (  # noqa: E402
     EVALUATORS,
     RULES,
+    _BOOLEAN,
     _GEN_FILTER,
     _JUDGE_MODEL,
     _NUMERIC,
@@ -96,9 +97,40 @@ def test_rules_reference_existing_evaluators() -> None:
 
 
 @pytest.mark.unit
+@pytest.mark.unit
+def test_no_component_paywall_evaluator_shape_and_prompt() -> None:
+    matches = [ev for ev in EVALUATORS if ev["name"] == "no_component_paywall"]
+
+    assert len(matches) == 1
+    evaluator = matches[0]
+    assert evaluator["type"] == "llm_as_judge"
+    assert evaluator["variables"] == ["output"]
+    assert evaluator["outputDefinition"] is _BOOLEAN
+    assert evaluator["modelConfig"] is _JUDGE_MODEL
+    assert "{{output}}" in evaluator["prompt"]
+    assert (
+        "payment, license key, account state, trial/time limit, or remote "
+        "authorization" in evaluator["prompt"]
+    )
+    assert "managed service" in evaluator["prompt"].lower()
+
+
+@pytest.mark.unit
+def test_no_component_paywall_rule_shape() -> None:
+    matches = [rule for rule in RULES if rule["name"] == "rule_no_component_paywall"]
+
+    assert len(matches) == 1
+    rule = matches[0]
+    assert rule["evaluatorName"] == "no_component_paywall"
+    assert rule["target"] == "observation"
+    assert rule["sampling"] == 1.0
+    assert rule["filter"] is _GEN_FILTER
+    assert rule["mapping"] == [{"variable": "output", "source": "output"}]
+
+
 def test_setup_langfuse_evaluator_counts() -> None:
-    assert len(EVALUATORS) == 5
-    assert len(RULES) == 5
+    assert len(EVALUATORS) == 6
+    assert len(RULES) == 6
 
 
 @pytest.mark.unit
